+++ tools/isobuild/package-source.js
@@ -766,13 +766,13 @@ _.extend(PackageSource.prototype, {
         implies: api.implies[arch],
         getFiles(sourceProcessorSet, watchSet) {
           const result = api.files[arch];
-          const relPathToSourceObj = {};
+          const relPathToSourceObj = Object.create(null);
           const sources = result.sources;
 
           // Files explicitly passed to api.addFiles remain at the
           // beginning of api.files[arch].sources in their given order.
-          sources.forEach(sourceObj => {
-            relPathToSourceObj[sourceObj.relPath] = sourceObj;
+          sources.forEach(source => {
+            relPathToSourceObj[source.relPath] = source;
           });
 
           self._findSources({
@@ -781,11 +781,21 @@ _.extend(PackageSource.prototype, {
             sourceArch: this,
             isApp: false
           }).forEach(relPath => {
-            if (! _.has(relPathToSourceObj, relPath)) {
-              const fileOptions = self._inferFileOptions(relPath, {
-                arch,
-                isApp: false,
-              });
+            const source = relPathToSourceObj[relPath];
+
+            if (source) {
+              const fileOptions = source.fileOptions ||
+                (source.fileOptions = Object.create(null));
+
+              // Since this file was explicitly added with api.addFiles or
+              // api.mainModule, it should be loaded eagerly unless the
+              // caller specified a boolean fileOptions.lazy value.
+              if (typeof fileOptions.lazy !== "boolean") {
+                fileOptions.lazy = false;
+              }
+
+            } else {
+              const fileOptions = Object.create(null);
 
               // Since this file was not explicitly added with
               // api.addFiles, it should not be evaluated eagerly.
@@ -838,8 +848,13 @@ _.extend(PackageSource.prototype, {
     });
 
     const projectWatchSet = projectContext.getProjectWatchSet();
+
     const mainModulesByArch =
       projectContext.meteorConfig.getMainModulesByArch();
+
+    const testModulesByArch =
+      projectContext.meteorConfig.getTestModulesByArch();
+
     projectWatchSet.merge(projectContext.meteorConfig.watchSet);
 
     _.each(compiler.ALL_ARCHES, function (arch) {
@@ -852,6 +867,9 @@ _.extend(PackageSource.prototype, {
 
       const mainModule = projectContext.meteorConfig
         .getMainModuleForArch(arch, mainModulesByArch);
+
+      const testModule = projectContext.meteorConfig
+        .getTestModuleForArch(arch, testModulesByArch);
 
       // XXX what about /web.browser/* etc, these directories could also
       // be for specific client targets.
@@ -885,10 +903,10 @@ _.extend(PackageSource.prototype, {
               missingMainModule = false;
             }
 
-            const fileOptions = self._inferFileOptions(relPath, {
+            const fileOptions = self._inferAppFileOptions(relPath, {
               arch,
-              isApp: true,
               mainModule,
+              testModule,
             });
 
             return {
@@ -947,22 +965,50 @@ _.extend(PackageSource.prototype, {
     }
   }),
 
-  _inferFileOptions(relPath, {
+  _inferAppFileOptions(relPath, {
     arch,
-    isApp,
     mainModule,
+    testModule,
   }) {
-    const fileOptions = {};
-    const isTest = global.testCommandMetadata
-      && global.testCommandMetadata.isTest;
-    const isAppTest = global.testCommandMetadata
-      && global.testCommandMetadata.isAppTest;
-    const isTestFile = (isTest || isAppTest) && isTestFilePath(relPath);
+    const fileOptions = Object.create(null);
+    const {
+      isTest = false,
+      isAppTest = false,
+    } = global.testCommandMetadata || {};
+
+    let isTestFile = false;
+    if (isTest || isAppTest) {
+      if (typeof testModule === "undefined") {
+        // If a testModule was not configured in the "meteor" section of
+        // package.json for this architecture, then isTestFilePath should
+        // determine whether this file loads eagerly.
+        isTestFile = isTestFilePath(relPath);
+      } else if (relPath === testModule) {
+        // If testModule is a string === relPath, then it is the entry
+        // point for tests, and should be loaded eagerly.
+        isTestFile = true;
+        fileOptions.lazy = false;
+        fileOptions.testModule = true;
+      } else {
+        // If testModule was defined but !== relPath, this file should not
+        // be loaded eagerly during tests. Setting fileOptions.testModule
+        // to false indicates that a testModule was configured, but this
+        // was not it. ResourceSlot#_isLazy (in compiler-plugin.js) will
+        // use this information (together with fileOptions.mainModule) to
+        // make the final call as to whether this file should be loaded
+        // eagerly or lazily.
+        isTestFile = false;
+        fileOptions.testModule = false;
+      }
+    }
 
     // If running in test mode (`meteor test`), all files other than
     // test files should be loaded lazily.
-    if (isTest && !isTestFile) {
+    if (isTest && ! isTestFile) {
       fileOptions.lazy = true;
+      // Ignore any meteor.mainModule that was specified, since we're
+      // running `meteor test` without the --full-app option.
+      mainModule = void 0;
     }
 
     const dirs = files.pathDirname(relPath).split(files.pathSep);
@@ -980,7 +1026,7 @@ _.extend(PackageSource.prototype, {
       }
 
       // Files in `imports/` should be lazily loaded *apart* from tests
-      if (isApp && dir === "imports" && !isTestFile) {
+      if (dir === "imports" && ! isTestFile) {
         fileOptions.lazy = true;
       }
 
@@ -999,18 +1045,23 @@ _.extend(PackageSource.prototype, {
       if (i > 0 &&
           dirs[i - 1] === "client" &&
           dir === "compatibility" &&
-          isApp && // Skip this check for packages.
           archinfo.matches(arch, "web") &&
           relPath.endsWith(".js")) {
         fileOptions.bare = true;
       }
     }
 
-    if (isApp && mainModule) {
+    if (typeof mainModule !== "undefined") {
+      // Note: if mainModule === false, no JavaScript modules will be
+      // loaded eagerly unless explicitly added with !fileOptions.lazy by
+      // a compiler plugin. This can be useful for building an app that
+      // does not run any application JS on the client (or the server). Of
+      // course, Meteor packages may still run JS on startup, but they
+      // have their own rules for lazy/eager loading of modules.
       if (relPath === mainModule) {
         fileOptions.lazy = false;
         fileOptions.mainModule = true;
-      } else if (typeof fileOptions.lazy === "undefined") {
+      } else {
         // Used in ResourceSlot#_isLazy (in compiler-plugin.js) to make a
         // final determination of whether the file should be lazy.
         fileOptions.mainModule = false;